fix implement MVC
